+++ ase/cell.py
@@ -205,7 +205,7 @@ class Cell:
 
         Equal to the number of nonzero lattice vectors."""
         # The name ndim clashes with ndarray.ndim
-        return sum(self.mask())  # type: ignore
+        return sum(self.mask())
 
     @property
     def orthorhombic(self) -> bool:
@@ -272,7 +272,7 @@ class Cell:
 
         Does not include factor of 2 pi."""
         icell = Cell(np.linalg.pinv(self).transpose())
-        icell[~self.mask()] = 0.0  # type: ignore
+        icell[~self.mask()] = 0.0  # type: ignore[index]
         return icell
 
     def normal(self, i):