+++ ase/data/g2_2.py
@@ -1776,8 +1776,8 @@ data = {
 # all constituent atoms
 atoms_g22 = []
 for f in data.keys():
-    s = Atoms(symbols=data[f]['symbols'],  # type: ignore
-              positions=data[f]['positions'])  # type: ignore
+    s = Atoms(symbols=data[f]['symbols'],  # type: ignore[index]
+              positions=data[f]['positions'])  # type: ignore[index]
     for a in s:
         atoms_g22.append(a.symbol)
 # unique atoms